requirements.txt
duckduckgo-search
langchain
openai
python-dotenv
streamlit
streamlit_chat
tiktoken
wikipedia